backend/requirements.txt
flask
flask_sqlalchemy
flask_restful
Flask-CORS
flask-security-too
celery[redis]
httplib2
jinja2
matplotlib
numpy
flask-caching